Create one line for hatch area of one polygon

Create one line for hatch area of one polygon

polygon.onehatch(
  x,
  y,
  x0,
  y0,
  xd,
  yd,
  ..debug.hatch = FALSE,
  fillOddEven = FALSE,
  ...
)

Arguments

x
vectors containing the coordinates of the vertices of the polygon.

y
vectors containing the coordinates of the vertices of the polygon.

x0
parameter as issued from polygon.fullhatch

y0
parameter as issued from polygon.fullhatch

xd
parameter as issued from polygon.fullhatch

yd
parameter as issued from polygon.fullhatch

..debug.hatch
for drawing when debugging function

fillOddEven
logical controlling the polygon shading mode: see below for details. Default FALSE.

...
graphical parameters such as xpd, lend, ljoin and lmitre can be given as arguments.